refactor: add comprehensive test suite for Visitor module, including CRUD and attendance functionality
